changed the scale of the game from 20 to 2, fixing an issue where bullet would not allow spheres to roll

+++ src/Map.hpp
@@ -79,7 +79,7 @@ class Map {
 		PrintType** mMap;
 		unsigned int mWidth;
 		unsigned int mHeight;
-		unsigned int mScale;
+		double mScale;
 		time_t mSeed;
 
 	private:
